+++ NEWS.md
@@ -2,6 +2,7 @@
 
   * Added getRatings method for topNList.
   * realRatingMatrix supports now subset assignment with [.
+  * RECOM_POPULAR now shows the parameters in the registry. 
 
 # Changes in version 0.1-9 (05/18/2016)